@@ -1,6 +1,6 @@
 {
   "name": "reactable",
-  "version": "0.14.1",
+  "version": "0.16.0",
   "description": "Fast, flexible, simple data tables in React",
   "main": "./lib/reactable.js",
   "repository": {
@@ -18,18 +18,23 @@
     "build": "./node_modules/grunt-cli/bin/grunt build"
   },
   "peerDependencies": {
-    "react": "* || ^0.14.0"
+    "react": "* || ^16.0.0",
+    "react-dom": "^16.0.0"
   },
   "devDependencies": {
+    "chai": "^4.1.2",
+    "chai-jquery": "^2.0.0",
+    "es6-shim": "^0.35.3",
     "grunt": "^0.4.4",
     "grunt-babel": "^5.0.1",
     "grunt-cli": "^0.1.13",
     "grunt-contrib-concat": "^0.5.1",
     "grunt-contrib-watch": "^0.6.1",
     "grunt-file-append": "0.0.6",
-    "grunt-karma": "^0.8.2",
+    "grunt-karma": "^2.0.0",
     "grunt-react": "^0.10.0",
-    "karma": "^0.12.9",
+    "jquery": "^3.3.1",
+    "karma": "^1.6.0",
     "karma-chrome-launcher": "^0.1.3",
     "karma-firefox-launcher": "^0.1.3",
     "karma-mocha": "^0.1.3",
@@ -38,7 +43,9 @@
     "load-grunt-tasks": "^3.2.0",
     "markdownlint": "0.0.7",
     "mocha": "^1.18.2",
-    "react": "^0.14.0"
+    "prop-types": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
   "keywords": [
     "react-component",
