@@ -1,64 +1,66 @@
 module.exports = function(config) {
-  config.set({
+    config.set({
 
-    // base path, that will be used to resolve files and exclude
-    basePath: '',
+        // base path, that will be used to resolve files and exclude
+        basePath: '',
 
-    // frameworks to use
-    frameworks: ['mocha'],
-
-
-    // list of files / patterns to load in the browser
-    files: [
-        'bower_components/es5-shim/es5-shim.js',
-        'bower_components/jquery/dist/jquery.js',
-        'bower_components/react/react-with-addons.js',
-        'bower_components/react/react-dom.js',
-        'bower_components/chai/chai.js',
-        'bower_components/chai-jquery/chai-jquery.js',
-        'build/reactable.js',
-        'build/tests/*.js',
-    ],
+        // frameworks to use
+        frameworks: ['mocha'],
 
 
-    // list of files to exclude
-    exclude: [ ],
+        // list of files / patterns to load in the browser
+        files: [
+            'node_modules/es6-shim/es6-shim.js',
+            'node_modules/jquery/dist/jquery.js',
+            'node_modules/react/umd/react.development.js',
+            'node_modules/react-dom/umd/react-dom.development.js',
+            'node_modules/react-dom/umd/react-dom-test-utils.development.js',
+            'node_modules/prop-types/prop-types.js',
+            'node_modules/chai/chai.js',
+            'node_modules/chai-jquery/chai-jquery.js',
+            'build/reactable.js',
+            'build/tests/*.js',
+        ],
 
 
-    // test results reporter to use
-    // possible values: 'dots', 'progress', 'junit', 'growl', 'coverage'
-    reporters: [ process.env.CI === 'true' ? 'spec' : 'dots'],
+        // list of files to exclude
+        exclude: [ ],
 
-    // web server port
-    port: 9876,
 
-    // enable / disable colors in the output (reporters and logs)
-    colors: true,
+        // test results reporter to use
+        // possible values: 'dots', 'progress', 'junit', 'growl', 'coverage'
+        reporters: [ process.env.CI === 'true' ? 'spec' : 'dots'],
 
-    // level of logging
-    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
-    logLevel: config.LOG_WARN,
+        // web server port
+        port: 9876,
 
-    // enable / disable watching file and executing tests whenever any file changes
-    autoWatch: false,
+        // enable / disable colors in the output (reporters and logs)
+        colors: true,
 
-    // Start these browsers, currently available:
-    // - Chrome
-    // - ChromeCanary
-    // - Firefox
-    // - Opera (has to be installed with `npm install karma-opera-launcher`)
-    // - Safari (only Mac; has to be installed with `npm install karma-safari-launcher`)
-    // - PhantomJS
-    // - IE (only Windows; has to be installed with `npm install karma-ie-launcher`)
-    browsers: ['PhantomJS'],
+        // level of logging
+        // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
+        logLevel: config.LOG_WARN,
 
-    // If browser does not capture in given timeout [ms], kill it
-    captureTimeout: 60000,
+        // enable / disable watching file and executing tests whenever any file changes
+        autoWatch: false,
 
-    // Continuous Integration mode
-    // if true, it capture browsers, run tests and exit
-    singleRun: true
-  });
+        // Start these browsers, currently available:
+        // - Chrome
+        // - ChromeCanary
+        // - Firefox
+        // - Opera (has to be installed with `npm install karma-opera-launcher`)
+        // - Safari (only Mac; has to be installed with `npm install karma-safari-launcher`)
+        // - PhantomJS
+        // - IE (only Windows; has to be installed with `npm install karma-ie-launcher`)
+        browsers: ['PhantomJS'],
+
+        // If browser does not capture in given timeout [ms], kill it
+        captureTimeout: 60000,
+
+        // Continuous Integration mode
+        // if true, it capture browsers, run tests and exit
+        singleRun: true
+    });
 };
 
 
