@@ -42,6 +42,10 @@ var _tfoot = require('./tfoot');
 
 var _paginator = require('./paginator');
 
+var _propTypes = require("prop-types");
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var Table = (function (_React$Component) {
     _inherits(Table, _React$Component);
 
@@ -840,12 +844,12 @@ Table.defaultProps = {
 };
 
 Table.propTypes = {
-    sortBy: _react2['default'].PropTypes.bool,
-    itemsPerPage: _react2['default'].PropTypes.number, // number of items to display per page
-    filterable: _react2['default'].PropTypes.array, // columns to look at when applying the filter specified by filterBy
-    filterBy: _react2['default'].PropTypes.string, // text to filter the results by (see filterable)
-    hideFilterInput: _react2['default'].PropTypes.bool, // Whether the default input field for the search/filter should be hidden or not
-    hideTableHeader: _react2['default'].PropTypes.bool, // Whether the table header should be hidden or not
-    noDataText: _react2['default'].PropTypes.string, // Text to be displayed in the event there is no data to show
-    noDataComponent: _react2['default'].PropTypes.func // function called to provide a component to display in the event there is no data to show (supercedes noDataText)
+    sortBy: _propTypes2['default'].bool,
+    itemsPerPage: _propTypes2['default'].number, // number of items to display per page
+    filterable: _propTypes2['default'].array, // columns to look at when applying the filter specified by filterBy
+    filterBy: _propTypes2['default'].string, // text to filter the results by (see filterable)
+    hideFilterInput: _propTypes2['default'].bool, // Whether the default input field for the search/filter should be hidden or not
+    hideTableHeader: _propTypes2['default'].bool, // Whether the table header should be hidden or not
+    noDataText: _propTypes2['default'].string, // Text to be displayed in the event there is no data to show
+    noDataComponent: _propTypes2['default'].func // function called to provide a component to display in the event there is no data to show (supercedes noDataText)
 };
